Make Beta the first stage in homebrew release (#21)
This index.ts script is called by our automations when releasing a new version to homebrew. The purpose is to generate a formula for GT Alpha with the correct download links to the GT binary for the new version.

Since we are dropping the Alpha track, we should make this script generate the formula for Beta directly.

I tested this script generation works in the upstack draft PR. I created some dummy files for different arch builds of GT to test the sha calculation and verified the materialized beta template is accurate.